@@ -20,56 +20,83 @@
   "keywords": [
     "react-component",
     "react",
-    "daterangepicker",
-    "datepicker"
+    "daterangepicker"
   ],
-  "author": "A2ZH",
+  "contributors": [
+    "A2ZH",
+    "Simon Guo <simonguo.2009@gmail.com>"
+  ],
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/rsuite/rsuite-daterangepicker/issues"
   },
   "homepage": "https://github.com/rsuite/rsuite-daterangepicker#readme",
+  "files": [
+    "CHANGELOG.md",
+    "lib"
+  ],
   "dependencies": {
-    "moment": "^2.17.1",
-    "classnames": "^2.2.5"
+    "classnames": "^2.2.5",
+    "dom-lib": "^0.2.6",
+    "lodash": "^4.17.4",
+    "moment": ">=2.15.2",
+    "prop-types": "^15.5.10",
+    "rsuite-theme": "^2.0.0-beta.1-1",
+    "rsuite-utils": "^0.0.9"
   },
   "peerDependencies": {
     "react": "^0.14.9 || >=15.3.0",
-    "react-dom": "^0.14.9 || >=15.3.0",
-    "rsuite": ">=1.0.0",
-    "dom-lib": ">=0.2.0"
+    "react-dom": "^0.14.9 || >=15.3.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.23.0",
-    "babel-core": "^6.23.1",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "^6.3.0",
-    "babel-preset-latest": "^6.22.0",
-    "babel-preset-react": "^6.23.0",
-    "babel-preset-stage-0": "^6.22.0",
-    "css-loader": "^0.26.2",
-    "dom-lib": "0.0.16",
-    "eslint": "^3.15.0",
+    "babel-core": "^6.18.0",
+    "babel-eslint": "^7.0.0",
+    "babel-loader": "^6.2.5",
+    "babel-plugin-istanbul": "^4.1.4",
+    "babel-preset-latest": "^6.16.0",
+    "babel-preset-react": "^6.11.1",
+    "babel-preset-stage-0": "^6.16.0",
+    "chai": "^3.5.0",
+    "coveralls": "^2.13.1",
+    "css-loader": "^0.25.0",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^15.0.1",
+    "eslint-plugin-babel": "^3.2.0",
+    "eslint-plugin-import": "^2.6.1",
+    "eslint-plugin-jsx-a11y": "^5.0.1",
+    "eslint-plugin-react": "^7.0.1",
     "extract-text-webpack-plugin": "^2.1.0",
+    "file-loader": "^0.10.1",
     "gh-pages": "^0.12.0",
-    "highlight.js": "^9.10.0",
     "html-loader": "^0.4.5",
-    "html-webpack-plugin": "^2.22.0",
-    "less": "^2.7.2",
-    "less-loader": "^3.0.0",
+    "html-webpack-plugin": "^2.28.0",
+    "jsdom": "^9.5.0",
+    "karma": "^1.1.1",
+    "karma-chrome-launcher": "^1.0.1",
+    "karma-cli": "^1.0.1",
+    "karma-coverage": "^1.1.0",
+    "karma-firefox-launcher": "^1.0.0",
+    "karma-mocha": "^1.1.1",
+    "karma-mocha-reporter": "^2.0.4",
+    "karma-sinon-chai": "^1.2.2",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-webpack": "2.0.3",
+    "less": "^2.7.1",
+    "less-loader": "^2.2.3",
     "markdown-loader": "2.0.0",
     "markdownloader": "1.0.5",
-    "marked": "^0.3.6",
-    "moment": "^2.17.1",
+    "mocha": "^3.1.0",
     "react": "^0.14.9",
     "react-dom": "^0.14.9",
     "react-hot-loader": "^3.0.0-beta.6",
-    "rsuite": "0.0.24",
-    "rsuite-theme": "0.3.24",
-    "style-loader": "^0.13.2",
+    "rsuite": "1.0.14",
+    "rsuite-affix": "^1.0.3",
+    "rsuite-theme": "^2.0.0-beta.1-1",
+    "sinon": "^1.17.4",
+    "sinon-chai": "^2.8.0",
+    "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
-    "file-loader": "^0.10.1",
-    "webpack": "^2.2.1",
+    "webpack": "^3.0.0",
     "webpack-dev-server": "^2.3.0"
   }
 }
