@@ -1,44 +1,32 @@
 const path = require('path');
 const webpack = require('webpack');
-
 const HtmlwebpackPlugin = require('html-webpack-plugin');
-const ExtractTextPlugin = require("extract-text-webpack-plugin");
-const marked = require("marked");
-const hl = require('highlight.js');
-const renderer = new marked.Renderer();
+const ExtractTextPlugin = require('extract-text-webpack-plugin');
+const markdownLoader = require('markdownloader').renderer;
 
-renderer.code = function (code, lang) {
-  lang = lang === 'js' ? 'javascript' : lang;
-  if (lang === 'html') {
-    lang = 'xml';
-  }
-  const hlCode = lang ? hl.highlight(lang, code).value : hl.highlightAuto(code).value;
-  return `<div class="doc-highlight"><pre><code class="${lang || ''}">${hlCode}</code></pre></div>`;
-};
 
 const extractLess = new ExtractTextPlugin({
-  filename: "[name].[contenthash].css",
-  disable: process.env.NODE_ENV === "development"
+  filename: '[name].[contenthash].css',
+  disable: process.env.NODE_ENV === 'development'
 });
 
 
 const docsPath = process.env.NODE_ENV === 'development' ? './assets' : './';
-
 const plugins = [
-  new webpack.HotModuleReplacementPlugin(),
-  new webpack.NamedModulesPlugin(),
   new webpack.DefinePlugin({
     'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV)
   }),
-  extractLess,
+  new webpack.HotModuleReplacementPlugin(),
+  new webpack.NamedModulesPlugin(),
   new HtmlwebpackPlugin({
-    title: 'RSUITE DateRangePicker',
+    title: 'RSUITE DatePicker',
     filename: 'index.html',
     template: 'docs/index.html',
     inject: true,
     hash: true,
     path: docsPath
-  })
+  }),
+  extractLess
 ];
 
 if (process.env.NODE_ENV === 'production') {
@@ -59,44 +47,52 @@ const common = {
   },
   plugins,
   module: {
-    rules: [{
-      test: /\.jsx?$/,
-      use: [
-        'babel-loader'
-      ],
-      exclude: /node_modules/
-    }, {
-      test: /\.less$/,
-      loader: extractLess.extract({
-        use: [{
-          loader: "css-loader"
-        }, {
-          loader: "less-loader"
-        }],
-        // use style-loader in development
-        fallback: "style-loader"
-      })
-    }, {
-      test: /\.md$/,
-      use: [{
-        loader: "html-loader"
+    rules: [
+      {
+        test: /\.jsx?$/,
+        use: [
+          'babel-loader'
+        ],
+        exclude: /node_modules/
+      },
+      {
+        test: /\.less$/,
+        loader: extractLess.extract({
+          use: [{
+            loader: 'css-loader'
+          }, {
+            loader: 'less-loader'
+          }],
+          // use style-loader in development
+          fallback: 'style-loader'
+        })
       }, {
-        loader: 'markdown-loader',
-        options: {
-          pedantic: true,
-          renderer
-        }
-      }]
-    }, {
-      test: /\.(woff|woff2|eot|ttf|svg)($|\?)/,
-      use: [{
-        loader: 'url-loader?limit=1&hash=sha512&digest=hex&size=16&name=resources/[hash].[ext]'
-      }]
-    }]
-  }
-}
+        test: /\.md$/,
+        use: [
+          {
+            loader: 'html-loader'
+          },
+          {
+            loader: 'markdown-loader',
+            options: {
+              pedantic: true,
+              renderer: markdownLoader.renderer
+            }
+          }
+        ]
+      },
+      {
+        test: /\.(woff|woff2|eot|ttf|svg)($|\?)/,
+        use: [{
+          loader: 'url-loader?limit=1&hash=sha512&digest=hex&size=16&name=resources/[hash].[ext]'
+        }]
 
-module.exports = (env = {}) => {
+      }
+    ]
+  }
+};
+
+module.exports = () => {
 
   if (process.env.NODE_ENV === 'development') {
     return Object.assign({}, common, {
@@ -110,11 +106,11 @@ module.exports = (env = {}) => {
     });
   }
 
-  if (process.env.NODE_ENV === 'production') {
-    return Object.assign({}, common, {
-      entry: [
-        path.resolve(__dirname, 'docs/index')
-      ]
-    });
-  }
-}
+
+  return Object.assign({}, common, {
+    entry: [
+      path.resolve(__dirname, 'docs/index')
+    ]
+  });
+
+};
